pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "dateparser>=1.2.1",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "imagehash>=4.3.2",
    "numpy>=2.3.0",
    "openai>=1.30.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uuid>=1.30",
    "uvicorn>=0.34.3",
]
